Restructure nav pages in NavbarLayout folder
